refactor: Voucher와 Customer 패키지 구조 분리
