+++ src/main/java/com/devcourse/voucherapp/CommandLineApplication.java
@@ -3,15 +3,15 @@ package com.devcourse.voucherapp;
 import com.devcourse.voucherapp.controller.CustomerController;
 import com.devcourse.voucherapp.controller.VoucherController;
 import com.devcourse.voucherapp.entity.Menu;
-import com.devcourse.voucherapp.entity.VoucherType;
-import com.devcourse.voucherapp.entity.dto.CustomerCreateRequestDto;
-import com.devcourse.voucherapp.entity.dto.CustomerResponseDto;
-import com.devcourse.voucherapp.entity.dto.CustomerUpdateRequestDto;
-import com.devcourse.voucherapp.entity.dto.CustomersResponseDto;
-import com.devcourse.voucherapp.entity.dto.VoucherCreateRequestDto;
-import com.devcourse.voucherapp.entity.dto.VoucherResponseDto;
-import com.devcourse.voucherapp.entity.dto.VoucherUpdateRequestDto;
-import com.devcourse.voucherapp.entity.dto.VouchersResponseDto;
+import com.devcourse.voucherapp.entity.voucher.VoucherType;
+import com.devcourse.voucherapp.entity.customer.request.CustomerCreateRequestDto;
+import com.devcourse.voucherapp.entity.customer.response.CustomerResponseDto;
+import com.devcourse.voucherapp.entity.customer.request.CustomerUpdateRequestDto;
+import com.devcourse.voucherapp.entity.customer.response.CustomersResponseDto;
+import com.devcourse.voucherapp.entity.voucher.request.VoucherCreateRequestDto;
+import com.devcourse.voucherapp.entity.voucher.response.VoucherResponseDto;
+import com.devcourse.voucherapp.entity.voucher.request.VoucherUpdateRequestDto;
+import com.devcourse.voucherapp.entity.voucher.response.VouchersResponseDto;
 import com.devcourse.voucherapp.view.ViewManager;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
