+++ src/main/java/com/devcourse/voucherapp/controller/VoucherController.java
@@ -1,9 +1,9 @@
 package com.devcourse.voucherapp.controller;
 
-import com.devcourse.voucherapp.entity.dto.VoucherCreateRequestDto;
-import com.devcourse.voucherapp.entity.dto.VoucherResponseDto;
-import com.devcourse.voucherapp.entity.dto.VoucherUpdateRequestDto;
-import com.devcourse.voucherapp.entity.dto.VouchersResponseDto;
+import com.devcourse.voucherapp.entity.voucher.request.VoucherCreateRequestDto;
+import com.devcourse.voucherapp.entity.voucher.response.VoucherResponseDto;
+import com.devcourse.voucherapp.entity.voucher.request.VoucherUpdateRequestDto;
+import com.devcourse.voucherapp.entity.voucher.response.VouchersResponseDto;
 import com.devcourse.voucherapp.service.VoucherService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Controller;
