+++ src/main/java/com/devcourse/voucherapp/entity/voucher/Voucher.java
@@ -1,6 +1,5 @@
 package com.devcourse.voucherapp.entity.voucher;
 
-import com.devcourse.voucherapp.entity.VoucherType;
 import java.util.UUID;
 
 public interface Voucher {
