+++ src/main/java/com/devcourse/voucherapp/repository/customer/CustomerRepository.java
@@ -1,4 +1,4 @@
-package com.devcourse.voucherapp.repository;
+package com.devcourse.voucherapp.repository.customer;
 
 import com.devcourse.voucherapp.entity.customer.Customer;
 import java.util.List;
