+++ src/main/java/com/devcourse/voucherapp/utils/exception/MenuInputException.java
@@ -1,4 +1,4 @@
-package com.devcourse.voucherapp.exception;
+package com.devcourse.voucherapp.utils.exception;
 
 import static java.text.MessageFormat.format;
